requirements/main.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --output-file requirements/main.txt requirements/main.in
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   starlette
    #   watchfiles
astropy==6.1.4
    # via
    #   -r requirements/main.in
    #   lsst-daf-butler
    #   lsst-felis
    #   lsst-resources
    #   lsst-utils
astropy-iers-data==0.2024.10.14.0.32.55
    # via astropy
backoff==2.2.1
    # via lsst-resources
boto3==1.35.42
    # via
    #   -r requirements/main.in
    #   lsst-resources
botocore==1.35.42
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   lsst-daf-butler
    #   lsst-dax-obscore
    #   lsst-felis
    #   safir
    #   uvicorn
cryptography==43.0.1
    # via
    #   pyjwt
    #   safir
defusedxml==0.7.1
    # via
    #   -r requirements/main.in
    #   lsst-resources
deprecated==1.2.14
    # via
    #   lsst-daf-butler
    #   lsst-daf-relation
    #   lsst-utils
fastapi==0.115.2
    # via
    #   -r requirements/main.in
    #   safir
gidgethub==5.3.0
    # via safir
google-api-core==2.21.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.35.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.18.2
    # via
    #   -r requirements/main.in
    #   lsst-resources
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.65.0
    # via google-api-core
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hpgeom==1.4.0
    # via lsst-sphgeom
httpcore==1.0.6
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.2
    # via
    #   lsst-daf-butler
    #   safir
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via -r requirements/main.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lsst-daf-butler==27.2024.4100
    # via
    #   -r requirements/main.in
    #   lsst-dax-obscore
lsst-daf-relation==27.2024.4100
    # via lsst-daf-butler
lsst-dax-obscore @ git+https://github.com/lsst-dm/dax_obscore@5919c78c29630b1c9ed7544a1e075d8cc7ebda22#egg=lsst-dax-obscore
    # via -r requirements/main.in
lsst-felis==27.2024.4100
    # via lsst-dax-obscore
lsst-resources==27.2024.4100
    # via
    #   -r requirements/main.in
    #   lsst-daf-butler
    #   lsst-dax-obscore
lsst-sphgeom==27.2024.3700
    # via
    #   lsst-daf-butler
    #   lsst-dax-obscore
lsst-utils==27.2024.4100
    # via
    #   lsst-daf-butler
    #   lsst-daf-relation
    #   lsst-dax-obscore
    #   lsst-felis
    #   lsst-resources
lxml==5.3.0
    # via pydantic-xml
markupsafe==3.0.1
    # via jinja2
numpy==2.1.2
    # via
    #   -r requirements/main.in
    #   astropy
    #   hpgeom
    #   lsst-daf-butler
    #   lsst-sphgeom
    #   lsst-utils
    #   pyarrow
    #   pyerfa
packaging==24.1
    # via astropy
proto-plus==1.24.0
    # via google-api-core
protobuf==5.28.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psutil==6.0.0
    # via lsst-utils
psycopg2==2.9.10
    # via lsst-daf-butler
psycopg2-binary==2.9.10
    # via -r requirements/main.in
pyarrow==17.0.0
    # via
    #   lsst-daf-butler
    #   lsst-dax-obscore
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   -r requirements/main.in
    #   fastapi
    #   lsst-daf-butler
    #   lsst-daf-relation
    #   lsst-felis
    #   pydantic-settings
    #   pydantic-xml
    #   safir
    #   vo-models
pydantic-core==2.23.4
    # via
    #   pydantic
    #   pydantic-xml
    #   safir
pydantic-settings==2.5.2
    # via -r requirements/main.in
pydantic-xml==2.13.1
    # via vo-models
pyerfa==2.0.1.4
    # via astropy
pyjwt==2.9.0
    # via gidgethub
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.12
    # via -r requirements/main.in
pyyaml==6.0.2
    # via
    #   astropy
    #   lsst-daf-butler
    #   lsst-dax-obscore
    #   lsst-felis
    #   lsst-utils
    #   uvicorn
requests==2.32.3
    # via
    #   -r requirements/main.in
    #   google-api-core
    #   google-cloud-storage
    #   lsst-resources
rsa==4.9
    # via google-auth
s3transfer==0.10.3
    # via boto3
safir==6.4.0
    # via -r requirements/main.in
safir-logging==6.4.0
    # via safir
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.36
    # via
    #   lsst-daf-butler
    #   lsst-daf-relation
    #   lsst-dax-obscore
    #   lsst-felis
starlette==0.40.0
    # via
    #   -r requirements/main.in
    #   fastapi
    #   safir
structlog==24.4.0
    # via
    #   safir
    #   safir-logging
threadpoolctl==3.5.0
    # via lsst-utils
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uritemplate==4.1.1
    # via gidgethub
urllib3==2.2.3
    # via
    #   botocore
    #   lsst-resources
    #   requests
uvicorn==0.32.0
    # via -r requirements/main.in
uvloop==0.21.0
    # via uvicorn
vo-models==0.4.1
    # via -r requirements/main.in
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
wrapt==1.16.0
    # via deprecated
